+++ src/main/java/chess/service/ChessService.java
@@ -9,7 +9,7 @@ import chess.controller.web.dto.game.GameRequestDto;
 import chess.controller.web.dto.history.HistoryResponseDto;
 import chess.controller.web.dto.score.ScoreResponseDto;
 import chess.controller.web.dto.state.StateResponseDto;
-import chess.dao.ChessDao;
+import chess.dao.*;
 import chess.domain.board.Board;
 import chess.domain.board.position.Path;
 import chess.domain.board.position.Position;
@@ -26,40 +26,48 @@ public class ChessService {
 
     private static final String MOVE_COMMAND_FORMAT = "move %s %s";
 
-    private final ChessDao chessDao;
+    private final GameDao gameDao;
+    private final HistoryDao historyDao;
+    private final PieceDao pieceDao;
+    private final ScoreDao scoreDao;
+    private final StateDao stateDao;
 
     public ChessService() {
-        this.chessDao = new ChessDao();
+        this.gameDao = new GameDao();
+        this.historyDao = new HistoryDao();
+        this.pieceDao = new PieceDao();
+        this.scoreDao = new ScoreDao();
+        this.stateDao = new StateDao();
     }
 
     public Long saveGame(final GameRequestDto gameRequestDto) {
         ChessManager chessManager = new ChessManager();
         Map<Position, Piece> pieces = chessManager.boardToMap();
-        Long gameId = chessDao.saveGame(gameRequestDto);
-        chessDao.saveState(chessManager, gameId);
-        chessDao.saveScore(chessManager.gameStatus(), gameId);
-        chessDao.savePieces(gameId, pieces);
+        Long gameId = gameDao.saveGame(gameRequestDto);
+        stateDao.saveState(chessManager, gameId);
+        scoreDao.saveScore(chessManager.gameStatus(), gameId);
+        pieceDao.savePieces(gameId, pieces);
         return gameId;
     }
 
-    public List<PieceResponseDto> findPiecesByGameId(final Long gameId) {
-        return chessDao.findPiecesByGameId(gameId);
+    public List<PieceResponseDto> findPiecesById(final Long gameId) {
+        return pieceDao.findPiecesByGameId(gameId);
     }
 
     public GameResponseDto findGameByGameId(final Long gameId) {
-        return chessDao.findGameByGameId(gameId);
+        return gameDao.findGameById(gameId);
     }
 
     public ScoreResponseDto findScoreByGameId(final Long gameId) {
-        return chessDao.findScoreByGameId(gameId);
+        return scoreDao.findScoreByGameId(gameId);
     }
 
     public StateResponseDto findStateByGameId(final Long gameId) {
-        return chessDao.findStateByGameId(gameId);
+        return stateDao.findStateByGameId(gameId);
     }
 
     public List<HistoryResponseDto> findHistoryByGameId(final Long gameId) {
-        return chessDao.findHistoryByGameId(gameId);
+        return historyDao.findHistoryByGameId(gameId);
     }
 
     public PathResponseDto movablePath(final MovablePathRequestDto movablePathRequestDto, final Long gameId) {
@@ -74,17 +82,17 @@ public class ChessService {
         HistoryResponseDto historyResponseDto = HistoryResponseDto.toChessManager(moveCommand, chessManager);
         Piece sourcePiece = chessManager.pickPiece(Position.of(moveRequestDto.getSource()));
         chessManager.move(Position.of(moveRequestDto.getSource()), Position.of(moveRequestDto.getTarget()));
-        chessDao.updateScore(chessManager.gameStatus(), gameId);
-        chessDao.updateState(chessManager, gameId);
-        chessDao.updateTargetPiece(moveRequestDto.getTarget(), sourcePiece, gameId);
-        chessDao.updateSourcePiece(moveRequestDto.getSource(), gameId);
-        chessDao.createHistory(historyResponseDto, gameId);
+        scoreDao.updateScore(chessManager.gameStatus(), gameId);
+        stateDao.updateState(chessManager, gameId);
+        pieceDao.updateTargetPiece(moveRequestDto.getTarget(), sourcePiece, gameId);
+        pieceDao.updateSourcePiece(moveRequestDto.getSource(), gameId);
+        historyDao.saveHistory(historyResponseDto, gameId);
         return historyResponseDto;
     }
 
     private ChessManager loadChessManager(final Long gameId) {
-        List<PieceResponseDto> pieceResponseDtos = chessDao.findPiecesByGameId(gameId);
-        StateResponseDto stateResponseDto = chessDao.findStateByGameId(gameId);
+        List<PieceResponseDto> pieceResponseDtos = pieceDao.findPiecesByGameId(gameId);
+        StateResponseDto stateResponseDto = stateDao.findStateByGameId(gameId);
         Map<Position, Piece> pieces = new HashMap<>();
         for (PieceResponseDto pieceResponseDto : pieceResponseDtos) {
             pieces.put(Position.of(pieceResponseDto.getPosition()), PieceConverter.parsePiece(pieceResponseDto.getSymbol()));
